It is now possible to create a meeting, with all of the required fields found on the Domain object.
The date field uses the ISO 8601 format as also used by XML. I had to write a RenderCustomJson result class (a modified version of the play.mvc.results.RenderJson class) in order to modify the Gson object and set the date format. I also created a RenderJSONEnhancer class which overrides the RenderJSON methods in play.mvc.Controller to call my RenderCustomJson instead of RenderJson. The date is now serialised and de-serialised in this format.

When creating a meeting, the owner of the meeting is set by the authorised user - as opposed to being read from JSON. This is a security precaution.