Added a custom LocationEventStream class based on the EventStream and ArchivedEventStream classes from play.libs.F
The purpose of this class is to (unlike the ArchivedEventStream) be able to remove a pipedStream (that is a currently connected stream that it holds reference to, for publishing purposes).

The change in functionality is that every location event is now published to a 'master' stream for each user. When a user connects, they obtain a stream which is a copy of the master and the master creates a reference to the connected stream. As they obtain a copy of all the messages in the master, they get a backlog of the previous messages as 'new' events (for the newly created stream). This is unlike before, when only new events would be published to a reconnected client.

Work still needs to be done on persisting the 'master' stream (archived stream) so that it can survive application restarts. Work will also need to be done in the future to enable multiple nodes and messaging of location events between nodes (as users could be connected to different nodes).